+++ src/lib/aloha/command.js
@@ -65,6 +65,10 @@ function( Aloha, Registry, Engine, Dom, ContentHandlerManager ) {
 	var CommandManager = {
 			
 		execCommand: function( commandId, showUi, value, range ) {
+			// native command fallback
+			if (Aloha.settings.nativeCommands === true) {
+				return document.execCommand(commandId, showUi, value);
+			}
 			
 			// Read current selection if not passed
 			if ( !range ) {
@@ -105,7 +109,12 @@ function( Aloha, Registry, Engine, Dom, ContentHandlerManager ) {
 		// If command is available and not disabled or the active range is not null 
 		// the command is enabled
 		queryCommandEnabled: function( commandId, range ) {
+			// native command fallback
+			if (Aloha.settings.nativeCommands === true) {
+				return document.queryCommandEnabled(commandId);
+			}
 
+			
 			// Take current selection if not passed
 			if ( !range ) {
 				if ( !Aloha.getSelection().getRangeCount() ) {
@@ -118,7 +127,11 @@ function( Aloha, Registry, Engine, Dom, ContentHandlerManager ) {
 
 		// "Return true if command is indeterminate, otherwise false."
 		queryCommandIndeterm: function( commandId, range ) {
-
+			// native command fallback
+			if (Aloha.settings.nativeCommands === true) {
+				return document.queryCommandIndeterm(commandId);
+			}
+			
 			// Take current selection if not passed
 			if ( !range ) {
 				if ( !Aloha.getSelection().getRangeCount() ) {
@@ -131,7 +144,11 @@ function( Aloha, Registry, Engine, Dom, ContentHandlerManager ) {
 		},
 		
 		queryCommandState: function( commandId, range ) {
-
+			// native command fallback
+			if (Aloha.settings.nativeCommands === true) {
+				return document.queryCommandState(commandId);
+			}
+			
 			// Take current selection if not passed
 			if ( !range ) {
 				if ( !Aloha.getSelection().getRangeCount() ) {
@@ -147,11 +164,19 @@ function( Aloha, Registry, Engine, Dom, ContentHandlerManager ) {
 		// interface is invoked, the user agent must return true if command is
 		// supported, and false otherwise."
 		queryCommandSupported: function( commandId ) {
-
+			// native command fallback
+			if (Aloha.settings.nativeCommands === true) {
+				return document.queryCommandSupported(commandId);
+			}
+			
 			return Engine.queryCommandSupported( commandId );		
 		},
 		
 		queryCommandValue: function( commandId, range ) {
+			// native command fallback
+			if (Aloha.settings.nativeCommands === true) {
+				return document.queryCommandValue(commandId);
+			}
 
 			// Take current selection if not passed
 			if ( !range ) {